pyproject.toml
[project]
name = "filmoclub"
version = "0.0.1"
description = "Cinephiles with tea"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "markdown>=3.7",
    "django-filter>=24.3",
    "pendulum>=3.0.0",
    "httpx[http2]>=0.27.2",
    "anyio[trio]>=4.6.1",
    "diskcache>=5.6.3",
    "django-cors-headers>=4.5.0",
    "pydantic>=2.9.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "icecream>=2.1.3",
    "cowsay>=6.1",
]
